// src/types/Profile/PersonalData.ts
export interface IProfileData {
  id: number;
  username: string;
  email: string;
  status: number;
  created_at: number;
  updated_at: number;
  fio: string;
  fio2: any;
  name: string;
  second_name: string;
  email_key: string;
  part: any;
  id_city: string;
  adres: string;
  tel: string;
  org: string;
  prim: string;
  ban: number;
  lat: any;
  lng: any;
  naim_city: string;
  geo: string;
  vzros: number;
  id_sms: any;
  sms: any;
  post_pass: any;
  id_city2: any;
  naim_city2: string;
  ref: any;
  urlfrom: string;
  refcount: any;
  refproc: any;
  showrek: any;
  id_refvidat: any;
  nom_refvidat: string;
  showrefurl: any;
  fiz_ur: any;
  id_vopl: number;
  inn: string;
  user_type: number;
  birthday: string;
  male: any;
  position: string;
  ind_pred: any;
  bank: any;
  schet: any;
  bik: any;
  id_firebase: any;
}